Neighborhood Search

TrixiParticles.GridNeighborhoodSearch — Type

GridNeighborhoodSearch{NDIMS}(search_radius, n_particles; periodic_box_min_corner=nothing,
                              periodic_box_max_corner=nothing, threaded_nhs_update=true)

Simple grid-based neighborhood search with uniform search radius. The domain is divided into a regular grid. For each (non-empty) grid cell, a list of particles in this cell is stored. Instead of representing a finite domain by an array of cells, a potentially infinite domain is represented by storing cell lists in a hash table (using Julia's Dict data structure), indexed by the cell index tuple

\[\left( \left\lfloor \frac{x}{d} \right\rfloor, \left\lfloor \frac{y}{d} \right\rfloor \right) \quad \text{or} \quad \left( \left\lfloor \frac{x}{d} \right\rfloor, \left\lfloor \frac{y}{d} \right\rfloor, \left\lfloor \frac{z}{d} \right\rfloor \right),\]

where $x, y, z$ are the space coordinates and $d$ is the search radius.

To find particles within the search radius around a point, only particles in the neighboring cells are considered.

See also (Chalela et al., 2021), (Ihmsen et al. 2011, Section 4.4).

As opposed to (Ihmsen et al. 2011), we do not sort the particles in any way, since not sorting makes our implementation a lot faster (although less parallelizable).

Arguments

NDIMS: Number of dimensions.

search_radius: The uniform search radius.

n_particles: Total number of particles.

Keywords

periodic_box_min_corner: In order to use a (rectangular) periodic domain, pass the coordinates of the domain corner in negative coordinate directions.

periodic_box_max_corner: In order to use a (rectangular) periodic domain, pass the coordinates of the domain corner in positive coordinate directions.

threaded_nhs_update=true: Can be used to deactivate thread parallelization in the neighborhood search update. This can be one of the largest sources of variations between simulations with different thread numbers due to particle ordering changes.

Internal use only

Please note that this constructor is intended for internal use only. It is not part of the public API of TrixiParticles.jl, and it thus can altered (or be removed) at any time without it being considered a breaking change.

To run a simulation with this neighborhood search, just pass the type to the constructor of Semidiscretization:

semi = Semidiscretization(system1, system2,
                          neighborhood_search=GridNeighborhoodSearch)

The keyword arguments periodic_box_min_corner and periodic_box_max_corner explained above can also be passed to the Semidiscretization and will internally be forwarded to the neighborhood search:

semi = Semidiscretization(system1, system2,
                          neighborhood_search=GridNeighborhoodSearch,
                          periodic_box_min_corner=[0.0, -0.25],
                          periodic_box_max_corner=[1.0, 0.75])

References

M. Chalela, E. Sillero, L. Pereyra, M.A. Garcia, J.B. Cabral, M. Lares, M. Merchán. "GriSPy: A Python package for fixed-radius nearest neighbors search". In: Astronomy and Computing 34 (2021). doi: 10.1016/j.ascom.2020.100443

Markus Ihmsen, Nadir Akinci, Markus Becker, Matthias Teschner. "A Parallel SPH Implementation on Multi-Core CPUs". In: Computer Graphics Forum 30.1 (2011), pages 99–112. doi: 10.1111/J.1467-8659.2010.01832.X

TrixiParticles.TrivialNeighborhoodSearch — Type

TrivialNeighborhoodSearch{NDIMS}(search_radius, eachparticle)

Trivial neighborhood search that simply loops over all particles. The search radius still needs to be passed in order to sort out particles outside the search radius in the internal function for_particle_neighbor, but it's not used in the internal function eachneighbor.

Arguments

NDIMS: Number of dimensions.

search_radius: The uniform search radius.

eachparticle: UnitRange of all particle indices. Usually just 1:n_particles.

Keywords

periodic_box_min_corner: In order to use a (rectangular) periodic domain, pass the coordinates of the domain corner in negative coordinate directions.

periodic_box_max_corner: In order to use a (rectangular) periodic domain, pass the coordinates of the domain corner in positive coordinate directions.

Internal use only

Please note that this constructor is intended for internal use only. It is not part of the public API of TrixiParticles.jl, and it thus can altered (or be removed) at any time without it being considered a breaking change.

To run a simulation with this neighborhood search, just pass the type to the constructor of Semidiscretization:

semi = Semidiscretization(system1, system2,
                          neighborhood_search=TrivialNeighborhoodSearch)

The keyword arguments periodic_box_min_corner and periodic_box_max_corner explained above can also be passed to the Semidiscretization and will internally be forwarded to the neighborhood search:

semi = Semidiscretization(system1, system2,
                          neighborhood_search=TrivialNeighborhoodSearch,
                          periodic_box_min_corner=[0.0, -0.25],
                          periodic_box_max_corner=[1.0, 0.75])