TODO

TrixiParticles.WindingNumberHorman — Type

WindingNumberHorman()

Algorithm for inside-outside segmentation of a complex geometry proposed by Horman et al. (2001). It is only supported for 2D geometries. WindingNumberHorman might handle edge cases a bit better, since the winding number is an integer value. Also, it is faster than WindingNumberJacobson for 2D geometries with more than about 100 edges.

Experimental Implementation

This is an experimental feature and may change in any future releases.

TrixiParticles.WindingNumberJacobson — Type

WindingNumberJacobson(; geometry=nothing, winding_number_factor=sqrt(eps()),
                      hierarchical_winding=false)

Algorithm for inside-outside segmentation of a complex geometry proposed by Jacobson et al. (2013).

Keywords

geometry: Complex geometry returned by load_geometry and is only required when using hierarchical_winding=true.

hierarchical_winding: If set to true, an optimized hierarchical approach will be used, which gives a significant speedup.

winding_number_factor: For leaky geometries, a factor of 0.4 will give a better inside-outside segmentation.

Experimental Implementation

This is an experimental feature and may change in any future releases.

TrixiParticles.load_geometry — Method

load_geometry(filename; element_type=Float64)

Load file and return corresponding type for ComplexShape. Supported file formats are .stl and .asc.

Arguments

filename: Name of the file to be loaded.

Keywords

element_type: Element type (default is Float64)

References

Alec Jacobson, Ladislav Kavan, and Olga Sorkine-Hornung "Robust inside-outside segmentation using generalized winding numbers". In: ACM Transactions on Graphics, 32.4 (2013), pages 1–12. doi: 10.1145/2461912.2461916

Kai Horman, Alexander Agathos "The point in polygon problem for arbitrary polygons". In: Computational Geometry, 20.3 (2001), pages 131–144. doi: 10.1016/s0925-7721(01)00012-8