Boundary System

TrixiParticles.WallBoundarySystem — Type

WallBoundarySystem(initial_condition, boundary_model;
                   prescribed_motion=nothing, adhesion_coefficient=0.0)

System for boundaries modeled by boundary particles. The interaction between fluid and boundary particles is specified by the boundary model.

Arguments

initial_condition: Initial condition (see InitialCondition)

boundary_model: Boundary model (see Boundary Models)

Keyword Arguments

prescribed_motion: For moving boundaries, a PrescribedMotion can be passed.

adhesion_coefficient: Coefficient specifying the adhesion of a fluid to the surface. Note: currently it is assumed that all fluids have the same adhesion coefficient.

TrixiParticles.BoundaryDEMSystem — Type

BoundaryDEMSystem(initial_condition, normal_stiffness)

System for boundaries modeled by boundary particles. The interaction between fluid and boundary particles is specified by the boundary model.

Experimental Implementation

This is an experimental feature and may change in a future releases.

TrixiParticles.PrescribedMotion — Type

PrescribedMotion(movement_function, is_moving; moving_particles=nothing)

Arguments

movement_function: Function of (x, t) where x is an SVector of the initial particle position and t is the time, returning an SVector of $d$ dimensions for a $d$-dimensional problem containing the new particle position at time t.

is_moving: Function of t to determine in each timestep if the particles are moving or not. Its boolean return value determines if the neighborhood search will be updated.

Keyword Arguments

moving_particles: Indices of moving particles. Default is each particle in the system.

Examples

# Circular motion of particles for t < 1.5
movement_function(x, t) = x + SVector(cos(2pi * t), sin(2pi * t))
is_moving(t) = t < 1.5

motion = PrescribedMotion(movement_function, is_moving)

# Rotation around the origin
movement_function2(x, t) = SVector(cos(2pi * t) * x[1] - sin(2pi * t) * x[2],
                                   sin(2pi * t) * x[1] + cos(2pi * t) * x[2])
is_moving2(t) = true

motion2 = PrescribedMotion(movement_function2, is_moving2)

Boundary Models

Dummy Particles

Boundaries modeled as dummy particles, which are treated like fluid particles, but their positions and velocities are not evolved in time. Since the force towards the fluid should not change with the material density when used with a TotalLagrangianSPHSystem, the dummy particles need to have a mass corresponding to the fluid's rest density, which we call "hydrodynamic mass", as opposed to mass corresponding to the material density of a TotalLagrangianSPHSystem.

Here, initial_density and hydrodynamic_mass are vectors that contains the initial density and the hydrodynamic mass respectively for each boundary particle. Note that when used with SummationDensity (see below), this is only used to determine the element type and the number of boundary particles.

To establish a relationship between density and pressure, a state_equation has to be passed, which should be the same as for the adjacent fluid systems. To sum over neighboring particles, a smoothing_kernel and smoothing_length needs to be passed. This should be the same as for the adjacent fluid system with the largest smoothing length.

In the literature, this kind of boundary particles is referred to as "dummy particles" (Adami et al., 2012 and Valizadeh & Monaghan, 2015), "frozen fluid particles" (Akinci et al., 2012) or "dynamic boundaries Crespo et al., 2007. The key detail of this boundary condition and the only difference between the boundary models in these references is the way the density and pressure of boundary particles is computed.

Since boundary particles are treated like fluid particles, the force on fluid particle $a$ due to boundary particle $b$ is given by

\[f_{ab} = m_a m_b \left( \frac{p_a}{\rho_a^2} + \frac{p_b}{\rho_b^2} \right) \nabla_{r_a} W(\Vert r_a - r_b \Vert, h).\]

The quantities to be defined here are the density $\rho_b$ and pressure $p_b$ of the boundary particle $b$.

TrixiParticles.BoundaryModelDummyParticles — Type

BoundaryModelDummyParticles(initial_density, hydrodynamic_mass,
                            density_calculator, smoothing_kernel,
                            smoothing_length; viscosity=nothing,
                            state_equation=nothing, correction=nothing,
                            reference_particle_spacing=0.0)

Boundary model for WallBoundarySystem.

Arguments

initial_density: Vector holding the initial density of each boundary particle.

hydrodynamic_mass: Vector holding the "hydrodynamic mass" of each boundary particle. See description above for more information.

density_calculator: Strategy to compute the hydrodynamic density of the boundary particles. See description below for more information.

smoothing_kernel: Smoothing kernel should be the same as for the adjacent fluid system.

smoothing_length: Smoothing length should be the same as for the adjacent fluid system.

Keywords

state_equation: This should be the same as for the adjacent fluid system (see e.g. StateEquationCole).

correction: Correction method of the adjacent fluid system (see Corrections).

viscosity: Slip (default) or no-slip condition. See description below for further information.

reference_particle_spacing: The reference particle spacing used for weighting values at the boundary, which currently is only needed when using surface tension.

Examples

# Free-slip condition
boundary_model = BoundaryModelDummyParticles(densities, masses, AdamiPressureExtrapolation(),
                                             smoothing_kernel, smoothing_length)

# No-slip condition
boundary_model = BoundaryModelDummyParticles(densities, masses, AdamiPressureExtrapolation(),
                                             smoothing_kernel, smoothing_length,
                                             viscosity=ViscosityAdami(nu=1e-6))

Hydrodynamic density of dummy particles

We provide six options to compute the boundary density and pressure, determined by the density_calculator:

(Recommended) With AdamiPressureExtrapolation, the pressure is extrapolated from the pressure of the fluid according to Adami et al., 2012, and the density is obtained by applying the inverse of the state equation. This option usually yields the best results of the options listed here.

(Only relevant for FSI) With BernoulliPressureExtrapolation, the pressure is extrapolated from the pressure similar to the AdamiPressureExtrapolation, but a relative velocity-dependent pressure part is calculated between moving bodies and fluids, which increases the boundary pressure in areas prone to penetrations.

With SummationDensity, the density is calculated by summation over the neighboring particles, and the pressure is computed from the density with the state equation.

With ContinuityDensity, the density is integrated from the continuity equation, and the pressure is computed from the density with the state equation. Note that this causes a gap between fluid and boundary where the boundary is initialized without any contact to the fluid. This is due to overestimation of the boundary density as soon as the fluid comes in contact with boundary particles that initially did not have contact to the fluid. Therefore, in dam break simulations, there is a visible "step", even though the boundary is supposed to be flat. See also dual.sphysics.org/faq/#Q_13.

With PressureZeroing, the density is set to the reference density and the pressure is computed from the density with the state equation. This option is not recommended. The other options yield significantly better results.

With PressureMirroring, the density is set to the reference density. The pressure is not used. Instead, the fluid pressure is mirrored as boundary pressure in the momentum equation. This option is not recommended due to stability issues. See PressureMirroring for more details.

1. AdamiPressureExtrapolation

The pressure of the boundary particles is obtained by extrapolating the pressure of the fluid according to Adami et al., 2012. The pressure of a boundary particle $b$ is given by

\[p_b = \frac{\sum_f (p_f + \rho_f (\bm{g} - \bm{a}_b) \cdot \bm{r}_{bf}) W(\Vert r_{bf} \Vert, h)}{\sum_f W(\Vert r_{bf} \Vert, h)},\]

where the sum is over all fluid particles, $\rho_f$ and $p_f$ denote the density and pressure of fluid particle $f$, respectively, $r_{bf} = r_b - r_f$ denotes the difference of the coordinates of particles $b$ and $f$, $\bm{g}$ denotes the gravitational acceleration acting on the fluid, and $\bm{a}_b$ denotes the acceleration of the boundary particle $b$.

TrixiParticles.AdamiPressureExtrapolation — Type

AdamiPressureExtrapolation(; pressure_offset=0, allow_loop_flipping=true)

density_calculator for BoundaryModelDummyParticles.

Keywords

pressure_offset=0: Sometimes it is necessary to artificially increase the boundary pressure to prevent penetration, which is possible by increasing this value.

allow_loop_flipping=true: Allow to flip the loop order for the pressure extrapolation. Disable to prevent error variations between simulations with different numbers of threads. Usually, the first (multithreaded) loop is over the boundary particles and the second loop over the fluid neighbors. When the number of boundary particles is larger than ceil(0.5 * nthreads()) times the number of fluid particles, it is usually more efficient to flip the loop order and loop over the fluid particles first. The factor depends on the number of threads, as the flipped loop is not thread parallelizable. This can cause error variations between simulations with different numbers of threads.

2. BernoulliPressureExtrapolation

Identical to the pressure $p_b$ calculated via AdamiPressureExtrapolation, but it adds the dynamic pressure component of the Bernoulli equation:

\[p_b = \frac{\sum_f (p_f + \frac{1}{2} \, \rho_{\text{neighbor}} \left( \frac{ (\mathbf{v}_f - \mathbf{v}_{\text{body}}) \cdot (\mathbf{x}_f - \mathbf{x}_{\text{neighbor}}) }{ \left\| \mathbf{x}_f - \mathbf{x}_{\text{neighbor}} \right\| } \right)^2 \times \text{factor} +\rho_f (\bm{g} - \bm{a}_b) \cdot \bm{r}_{bf}) W(\Vert r_{bf} \Vert, h)}{\sum_f W(\Vert r_{bf} \Vert, h)}\]

where $\mathbf{v}_f$ is the velocity of the fluid and $\mathbf{v}_{\text{body}}$ is the velocity of the body. This adjustment provides a higher boundary pressure for bodies moving with a relative velocity to the fluid to prevent penetration. This modification is original and not derived from any literature source.

TrixiParticles.BernoulliPressureExtrapolation — Type

BernoulliPressureExtrapolation(; pressure_offset=0, factor=1)

density_calculator for BoundaryModelDummyParticles.

Keywords

pressure_offset=0: Sometimes it is necessary to artificially increase the boundary pressure to prevent penetration, which is possible by increasing this value.

factor=1 : Setting factor allows to just increase the strength of the dynamic pressure part.

allow_loop_flipping=true: Allow to flip the loop order for the pressure extrapolation. Disable to prevent error variations between simulations with different numbers of threads. Usually, the first (multithreaded) loop is over the boundary particles and the second loop over the fluid neighbors. When the number of boundary particles is larger than ceil(0.5 * nthreads()) times the number of fluid particles, it is usually more efficient to flip the loop order and loop over the fluid particles first. The factor depends on the number of threads, as the flipped loop is not thread parallelizable. This can cause error variations between simulations with different numbers of threads.

5. PressureZeroing

This is the simplest way to implement dummy boundary particles. The density of each particle is set to the reference density and the pressure to the reference pressure (the corresponding pressure to the reference density by the state equation).

TrixiParticles.PressureZeroing — Type

PressureZeroing()

density_calculator for BoundaryModelDummyParticles.

Note

This boundary model produces significantly worse results than all other models and is only included for research purposes.

6. PressureMirroring

Instead of calculating density and pressure for each boundary particle, we modify the momentum equation,

\[\frac{\mathrm{d}v_a}{\mathrm{d}t} = -\sum_b m_b \left( \frac{p_a}{\rho_a^2} + \frac{p_b}{\rho_b^2} \right) \nabla_a W_{ab}\]

to replace the unknown density $\rho_b$ if $b$ is a boundary particle by the reference density and the unknown pressure $p_b$ if $b$ is a boundary particle by the pressure $p_a$ of the interacting fluid particle. The momentum equation therefore becomes

\[\frac{\mathrm{d}v_a}{\mathrm{d}t} = -\sum_f m_f \left( \frac{p_a}{\rho_a^2} + \frac{p_f}{\rho_f^2} \right) \nabla_a W_{af} -\sum_b m_b \left( \frac{p_a}{\rho_a^2} + \frac{p_a}{\rho_0^2} \right) \nabla_a W_{ab},\]

where the first sum is over all fluid particles and the second over all boundary particles.

This approach was first mentioned by Akinci et al. (2012) and written down in this form by Band et al. (2018).

TrixiParticles.PressureMirroring — Type

PressureMirroring()

density_calculator for BoundaryModelDummyParticles.

Note

This boundary model requires high viscosity for stability with WCSPH. It also produces significantly worse results than AdamiPressureExtrapolation and is not more efficient because smaller time steps are required due to more noise in the pressure. We added this model only for research purposes and for comparison with SPlisHSPlasH.

No-slip conditions

For the interaction of dummy particles and fluid particles, Adami et al. (2012) impose a no-slip boundary condition by assigning a wall velocity $v_w$ to the dummy particle.

The wall velocity of particle $a$ is calculated from the prescribed boundary particle velocity $v_a$ and the smoothed velocity field

\[v_w = 2 v_a - \frac{\sum_b v_b W_{ab}}{\sum_b W_{ab}},\]

where the sum is over all fluid particles.

By choosing the viscosity model ViscosityAdami for viscosity, a no-slip condition is imposed. It is recommended to choose nu in the order of either the kinematic viscosity parameter of the adjacent fluid or the equivalent from the artificial parameter alpha of the adjacent fluid ($\nu = \frac{\alpha h c }{2d + 4}$). When omitting the viscous interaction (default viscosity=nothing), a free-slip wall boundary condition is applied.

Warning

The viscosity model ArtificialViscosityMonaghan for BoundaryModelDummyParticles has not been verified yet.

Repulsive Particles

Boundaries modeled as boundary particles which exert forces on the fluid particles (Monaghan, Kajtar, 2009). The force on fluid particle $a$ due to boundary particle $b$ is given by

\[f_{ab} = m_a \left(\tilde{f}_{ab} - m_b \Pi_{ab} \nabla_{r_a} W(\Vert r_a - r_b \Vert, h)\right)\]

with

\[\tilde{f}_{ab} = \frac{K}{\beta^{n-1}} \frac{r_{ab}}{\Vert r_{ab} \Vert (\Vert r_{ab} \Vert - d)} \Phi(\Vert r_{ab} \Vert, h) \frac{2 m_b}{m_a + m_b},\]

where $m_a$ and $m_b$ are the masses of fluid particle $a$ and boundary particle $b$ respectively, $r_{ab} = r_a - r_b$ is the difference of the coordinates of particles $a$ and $b$, $d$ denotes the boundary particle spacing and $n$ denotes the number of dimensions (see Monaghan & Kajtar, 2009, Equation (3.1) and Valizadeh & Monaghan, 2015). Note that the repulsive acceleration $\tilde{f}_{ab}$ does not depend on the masses of the boundary particles. Here, $\Phi$ denotes the 1D Wendland C4 kernel, normalized to $1.77$ for $q=0$ (Monaghan & Kajtar, 2009, Section 4), with $\Phi(r, h) = w(r/h)$ and

\[w(q) = \begin{cases} (1.77/32) (1 + (5/2)q + 2q^2)(2 - q)^5 & \text{if } 0 \leq q < 2 \\ 0 & \text{if } q \geq 2. \end{cases}\]

The boundary particles are assumed to have uniform spacing by the factor $\beta$ smaller than the expected fluid particle spacing. For example, if the fluid particles have an expected spacing of $0.3$ and the boundary particles have a uniform spacing of $0.1$, then this parameter should be set to $\beta = 3$. According to Monaghan & Kajtar (2009), a value of $\beta = 3$ for the Wendland C4 that we use here is reasonable for most computing purposes.

The parameter $K$ is used to scale the force exerted by the boundary particles. In Monaghan & Kajtar (2009), a value of $gD$ is used for static tank simulations, where $g$ is the gravitational acceleration and $D$ is the depth of the fluid.

The viscosity $\Pi_{ab}$ is calculated according to the viscosity used in the simulation, where the density of the boundary particle if needed is assumed to be identical to the density of the fluid particle.

No-slip condition

By choosing the viscosity model ArtificialViscosityMonaghan for viscosity, a no-slip condition is imposed. When omitting the viscous interaction (default viscosity=nothing), a free-slip wall boundary condition is applied.

Warning

The no-slip conditions for BoundaryModelMonaghanKajtar have not been verified yet.

TrixiParticles.BoundaryModelMonaghanKajtar — Type

BoundaryModelMonaghanKajtar(K, beta, boundary_particle_spacing, mass;
                            viscosity=nothing)

Boundary model for WallBoundarySystem.

Arguments

K: Scaling factor for repulsive force.

beta: Ratio of fluid particle spacing to boundary particle spacing.

boundary_particle_spacing: Boundary particle spacing.

mass: Vector holding the mass of each boundary particle.

Keywords

viscosity: Free-slip (default) or no-slip condition. See description above for further information.

Open Boundaries

TrixiParticles.OpenBoundarySystem — Type

OpenBoundarySystem(boundary_zone::BoundaryZone;
                   fluid_system::AbstractFluidSystem, buffer_size::Integer,
                   boundary_model)

Open boundary system for in- and outflow particles.

Arguments

boundary_zone: See BoundaryZone.

Keywords

fluid_system: The corresponding fluid system

boundary_model: Boundary model (see Open Boundary Models)

buffer_size: Number of buffer particles.

Experimental Implementation

This is an experimental feature and may change in any future releases.

TrixiParticles.BoundaryZone — Type

BoundaryZone(; boundary_face, face_normal, density, particle_spacing,
             initial_condition=nothing, extrude_geometry=nothing,
             open_boundary_layers::Integer, average_inflow_velocity=true,
             boundary_type=BidirectionalFlow(),
             reference_density=nothing, reference_pressure=nothing,
             reference_velocity=nothing)

Boundary zone for OpenBoundarySystem.

The specified boundary_face (line in 2D or rectangle in 3D) will be extruded in the direction opposite to face_normal to create a box for the boundary zone. To specify the boundary_face, pass the required vertices as described below. For complex 3D simulations, these vertices can also be extracted from an STL file (see planar_geometry_to_face). There are three ways to specify the actual shape of the boundary zone:

Don't pass initial_condition or extrude_geometry. The boundary zone box will then be filled with boundary particles (default).

Specify extrude_geometry by passing a 1D shape in 2D or a 2D shape in 3D, which is then extruded in the direction opposite to face_normal to create the boundary particles.

In 2D, the shape must be either an initial condition with 2D coordinates, which lies on the line specified by boundary_face, or an initial condition with 1D coordinates, which lies on the line specified by boundary_face when a y-coordinate of 0 is added.

In 3D, the shape must be either an initial condition with 3D coordinates, which lies in the rectangle specified by boundary_face, or an initial condition with 2D coordinates, which lies in the rectangle specified by boundary_face when a z-coordinate of 0 is added.

Specify initial_condition by passing a 2D initial condition in 2D or a 3D initial condition in 3D, which will be used for the boundary particles.

Note

Particles outside the boundary zone box will be removed.

Keywords

boundary_face: Tuple of vertices defining a part of the surface of the domain. The vertices must either span a line in 2D or a rectangle in 3D. This line or rectangle is then extruded in upstream direction to obtain the boundary zone. In 2D, pass two vertices $(A, B)$, so that the interval $[A, B]$ is the inflow surface. In 3D, pass three vertices $(A, B, C)$, so that the rectangular inflow surface is spanned by the vectors $\widehat{AB}$ and $\widehat{AC}$. These two vectors must be orthogonal.

face_normal: Vector defining the normal of the boundary_face. It always points inside the fluid domain.

boundary_type=BidirectionalFlow(): Specify the type of the boundary. Available types are

InFlow() for an inflow boundary

OutFlow() for an outflow boundary

BidirectionalFlow() (default) for an bidirectional flow boundary

open_boundary_layers: Number of particle layers in the direction opposite to face_normal.

particle_spacing: The spacing between the particles (see InitialCondition).

density: Particle density (see InitialCondition).

initial_condition=nothing: InitialCondition for the inflow particles. Particles outside the boundary zone will be removed. Do not use together with extrude_geometry.

extrude_geometry=nothing: 1D shape in 2D or 2D shape in 3D, which lies on the boundary_face and is extruded upstream to obtain the inflow particles. See point 2 above for more details.

average_inflow_velocity=true: If true, the extrapolated inflow velocity is averaged to impose a uniform inflow profile. When no velocity is prescribed at the inflow, the velocity is extrapolated from the fluid domain. Thus, turbulent flows near the inflow can lead to anisotropic buffer-particles distribution, resulting in a potential numerical instability. Averaging mitigates these effects.

reference_velocity: Reference velocity is either a function mapping each particle's coordinates and time to its velocity, or, for a constant fluid velocity, a vector holding this velocity.

reference_pressure: Reference pressure is either a function mapping each particle's coordinates and time to its pressure, or a scalar for a constant pressure over all particles.

reference_density: Reference density is either a function mapping each particle's coordinates and time to its density, or a scalar for a constant density over all particles.

Note

The reference values (reference_velocity, reference_pressure, reference_density) can also be set to nothing. In this case, they will either be extrapolated from the fluid domain (BoundaryModelMirroringTafuni) or evolved using the characteristic flow variables (BoundaryModelCharacteristicsLastiwka).

Examples

# 2D
face_vertices = ([0.0, 0.0], [0.0, 1.0])
face_normal = [1.0, 0.0]

# Constant reference velocity:
velocity_const = [1.0, 0.0]

inflow_1 = BoundaryZone(; boundary_face=face_vertices, face_normal, particle_spacing=0.1,
                        density=1.0, open_boundary_layers=4, boundary_type=InFlow(),
                        reference_velocity=velocity_const)

# Reference velocity as a function (parabolic velocity profile):
velocity_func = (pos, t) -> SVector(4.0 * pos[2] * (1.0 - pos[2]), 0.0)

inflow_2 = BoundaryZone(; boundary_face=face_vertices, face_normal, particle_spacing=0.1,
                        density=1.0, open_boundary_layers=4, boundary_type=InFlow(),
                        reference_velocity=velocity_func)

# 3D
face_vertices = ([0.0, 0.0, 0.0], [1.0, 0.0, 0.0], [0.0, 1.0, 0.0])
face_normal = [0.0, 0.0, 1.0]

# Constant reference pressure:
pressure_const = 0.0

outflow_1 = BoundaryZone(; boundary_face=face_vertices, face_normal, particle_spacing=0.1,
                         density=1.0, open_boundary_layers=4, boundary_type=OutFlow(),
                         reference_pressure=pressure_const)

# Reference pressure as a function (y-dependent profile, sinusoidal in time):
pressure_func = (pos, t) -> pos[2] * sin(2pi * t)

outflow_2 = BoundaryZone(; boundary_face=face_vertices, face_normal, particle_spacing=0.1,
                         density=1.0, open_boundary_layers=4, boundary_type=OutFlow(),
                         reference_pressure=pressure_func)

# 3D particles sampled as cylinder
circle = SphereShape(0.1, 0.5, (0.5, 0.5), 1.0, sphere_type=RoundSphere())

bidirectional_flow = BoundaryZone(; boundary_face=face_vertices, face_normal,
                                  particle_spacing=0.1, density=1.0,
                                  boundary_type=BidirectionalFlow(),
                                  extrude_geometry=circle, open_boundary_layers=4)

Experimental Implementation

This is an experimental feature and may change in any future releases.

TrixiParticles.planar_geometry_to_face — Method

planar_geometry_to_face(planar_geometry::TriangleMesh)

Extracts a simplified rectangular face and its normal vector from an arbitrary planar geometry (TriangleMesh loaded via load_geometry) for use as a boundary zone interface in BoundaryZone. This function computes the corner points of an oriented bounding box that best represents the essential orientation and extent of the input geometry. The geometry must be planar (all vertices should lie in the same plane), but can have complex or non-rectangular boundaries.

Face Normal Orientation

All face normals of the input geometry must point inside the fluid domain. The returned plane normal is computed by averaging all face normals, so consistent orientation is required.

Arguments

planar_geometry: A planar geometry (TriangleMesh loaded via load_geometry).

Returns

face_vertices: Tuple of three vertices defining the rectangular face (corner points of the oriented bounding box).

face_normal: Normalized normal vector of the face.

Example

file = pkgdir(TrixiParticles, "test", "preprocessing", "data")
planar_geometry = load_geometry(joinpath(file, "inflow_geometry.stl"))

face, face_normal = planar_geometry_to_face(planar_geometry)

Open Boundary Models

We offer two models for open boundaries, with the choice depending on the specific problem and flow characteristics near the boundary:

Method of characteristics: The method of characteristics is typically used in problems where tracking of wave propagation or flow in a domain that interacts with open boundaries (e.g., shock waves, wave fronts, or any behavior that depends on the direction of propagation) is needed. It avoids artificial reflections that could arise from boundary conditions.

Mirroring: The mirroring method is often applied when the flow near the boundary is expected to behave in a way that is easier to model by using symmetry or when the fluid does not exhibit complex wave behavior near the boundary (e.g., free-surface flows and simple outflow).

Method of characteristics

TrixiParticles.BoundaryModelCharacteristicsLastiwka — Type

BoundaryModelCharacteristicsLastiwka(; extrapolate_reference_values=nothing)

Boundary model for OpenBoundarySystem. This model uses the characteristic variables to propagate the appropriate values to the outlet or inlet and was proposed by Lastiwka et al. (2009). It requires a specific flow direction to be passed to the BoundaryZone. For more information about the method see description below.

Keywords

extrapolate_reference_values=nothing: If one of the following mirroring methods is selected, the reference values are extrapolated from the fluid domain to the boundary particles:

ZerothOrderMirroring

FirstOrderMirroring

SimpleMirroring

Note: This feature is experimental and has not been fully validated yet. As of now, we are not aware of any published literature supporting its use. Note that even without this extrapolation feature, the reference values don't need to be prescribed - they're computed from the characteristics.

The difficulty in non-reflecting boundary conditions, also called open boundaries, is to determine the appropriate boundary values of the exact characteristics of the Euler equations. Assuming the flow near the boundaries is normal to the boundary and free of shock waves and significant viscous effects, it can be shown that three characteristic variables exist:

$J_1$, associated with convection of entropy and propagates at flow velocity,

$J_2$, downstream-running characteristics,

$J_3$, upstream-running characteristics.

Giles (1990) derived those variables based on a linearized set of governing equations:

\[J_1 = -c_s^2 (\rho - \rho_{\text{ref}}) + (p - p_{\text{ref}})\]

\[J_2 = \rho c_s (v - v_{\text{ref}}) + (p - p_{\text{ref}})\]

\[J_3 = - \rho c_s (v - v_{\text{ref}}) + (p - p_{\text{ref}})\]

where the subscript "ref" denotes the reference flow near the boundaries, which can be prescribed.

Specifying the reference variables is not equivalent to prescription of $\rho$, $v$ and $p$ directly, since the perturbation from the reference flow is allowed.

Lastiwka et al. (2009) applied the method of characteristic to SPH and determined the number of variables that should be prescribed at the boundary and the number which should be propagated from the fluid domain to the boundary:

For an inflow boundary:

Prescribe downstream-running characteristics $J_1$ and $J_2$

Transmit $J_3$ from the fluid domain (allow $J_3$ to propagate upstream to the boundary).

For an outflow boundary:

Prescribe upstream-running characteristic $J_3$

Transmit $J_1$ and $J_2$ from the fluid domain.

Prescribing is done by simply setting the characteristics to zero. To transmit the characteristics from the fluid domain, or in other words, to carry the information of the fluid to the boundaries, Negi et al. (2020) use a Shepard Interpolation

\[f_i = \frac{\sum_j^N f_j W_{ij}}{\sum_j^N W_{ij}},\]

where the $i$-th particle is a boundary particle, $f$ is either $J_1$, $J_2$ or $J_3$ and $N$ is the set of neighboring fluid particles.

To express pressure $p$, density $\rho$ and velocity $v$ as functions of the characteristic variables, the system of equations from the characteristic variables is inverted and gives

\[ \rho - \rho_{\text{ref}} = \frac{1}{c_s^2} \left( -J_1 + \frac{1}{2} J_2 + \frac{1}{2} J_3 \right),\]

\[u - u_{\text{ref}}= \frac{1}{2\rho c_s} \left( J_2 - J_3 \right),\]

\[p - p_{\text{ref}} = \frac{1}{2} \left( J_2 + J_3 \right).\]

With $J_1$, $J_2$ and $J_3$ determined, we can easily solve for the actual variables for each particle.

Mirroring

TrixiParticles.BoundaryModelMirroringTafuni — Type

BoundaryModelMirroringTafuni(; mirror_method=FirstOrderMirroring())

Boundary model for the OpenBoundarySystem. This model implements the method of Tafuni et al. (2018) to extrapolate the properties from the fluid domain to the buffer zones (inflow and outflow) using ghost nodes. The position of the ghost nodes is obtained by mirroring the boundary particles into the fluid along a direction that is normal to the open boundary. Fluid properties are then interpolated at these ghost node positions using surrounding fluid particles. The values are then mirrored back to the boundary particles. We provide three different mirroring methods: - ZerothOrderMirroring: Uses a Shepard interpolation to interpolate the values. - FirstOrderMirroring: Uses a first order correction based on the gradient of the interpolated values . - SimpleMirroring: Similar to the first order mirroring, but does not use the gradient of the interpolated values.

TrixiParticles.FirstOrderMirroring — Type

FirstOrderMirroring(; firstorder_tolerance=1f-3)

Fluid properties are extrapolated onto ghost nodes using the method proposed by Liu and Liu (2006), to extend the gradient into the boundary zone.

Keywords

firstorder_tolerance=1f-3: If the determinant of the correction matrix is smaller than this value, the method falls back to ZerothOrderMirroring.

TrixiParticles.SimpleMirroring — Type

SimpleMirroring(; firstorder_tolerance=1f-3))

This method is similar to FirstOrderMirroring, but does not use the corrected gradient as proposed by Negi et al. (2022).

Keywords

firstorder_tolerance=1f-3: If the determinant of the correction matrix is smaller than this value, the method falls back to ZerothOrderMirroring.

TrixiParticles.ZerothOrderMirroring — Type

ZerothOrderMirroring()

Fluid properties are interpolated onto ghost nodes using Shepard interpolation. (See slide 6 from the 4th DualSPHysics Users Workshop: Tafuni, Lisbon 2018). The position of the ghost nodes is obtained by mirroring the boundary particles into the fluid along a direction that is normal to the open boundary. The interpolated values at the ghost nodes are then assigned to the corresponding boundary particles.