Discrete Element Method

The Discrete Element Method (DEM) is a computational technique widely used in physics, engineering, and applied mathematics for simulating the mechanical behavior of granular materials, such as powders, sand, soil, or rock, as well as other discontinua. Unlike continuum mechanics that treats materials as continuous, DEM considers individual particles or elements and their interactions. This approach provides detailed insights into the micro-mechanical behavior of materials, making it particularly valuable in fields such as geomechanics, material science, and mechanical engineering.

Fundamental Principles

The core idea behind DEM is the discretization of a material system into a finite set of distinct, interacting mass elements (particles). These elements (particles) can vary in shape, size, and properties, and they interact with each other and possibly with their boundaries through contact forces and potential fields. The motion and behavior of each mass element are governed by Newton's laws of motion, accounting for the forces and moments acting upon them.

TrixiParticles.DEMSystem — Type

DEMSystem(initial_condition, normal_stiffness, elastic_modulus, poissons_ratio;
 damping_coefficient=0.0001, acceleration=ntuple(_ -> 0.0, NDIMS), source_terms=nothing)

Constructs a Discrete Element Method (DEM) system for numerically simulating the dynamics of granular and particulate matter. DEM is employed to simulate and analyze the motion, interactions, and collective behavior of assemblies of discrete, solid particles, typically under mechanical loading. The model accounts for individual particle characteristics and implements interaction laws that govern contact forces (normal and tangential), based on specified material properties and contact mechanics.

Arguments

initial_condition: Initial condition of the system, encapsulating the initial positions, velocities, masses, and radii of particles.

normal_stiffness: Normal stiffness coefficient for particle-particle and particle-wall contacts.

elastic_modulus: Elastic modulus for this particle system.

poissons_ratio: Poisson ratio for this particle system.

Keywords

acceleration: Global acceleration vector applied to the system, such as gravity. Specified as an SVector of length NDIMS, with a default of zero in each dimension.

source_terms: Optional; additional forces or modifications to particle dynamics not captured by standard DEM interactions, such as electromagnetic forces or user-defined perturbations.

damping_coefficient=0.0001: Set a damping coefficient for the collision interactions.

!!! warning "Experimental Implementation" This is an experimental feature and may change in a future releases.

References

N. Bićanić. "Discrete element methods". In: Encyclopedia of Computational Mechanics (2007). doi: 10.1002/0470091355.ecm006.pub2

P. Cundall and O. Strack. "A discrete numerical model for granular assemblies". In: Géotechnique 29.1 (1979), pages 47–65. doi: 10.1680/geot.1979.29.1.47

A. Renzo and F. Maio. "Comparison of contact-force models for the simulation of collisions in DEM-based granular flow codes" In: Chemical Engineering Science 59.3 (2004), pages 525–541. doi: 10.1016/j.ces.2003.09.037